start screen changed, main drawer redesign
