+++ .dart_tool/package_config_subset
@@ -1,0 +1,113 @@
+catalogo_filmes
+2.16
+file:///home/hiltonthallyson/Documents/UFRN/Dev%20Mobile/projects/projeto02/catalogo-filmes/
+file:///home/hiltonthallyson/Documents/UFRN/Dev%20Mobile/projects/projeto02/catalogo-filmes/lib/
+async
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.8.2/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.8.2/lib/
+boolean_selector
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0/lib/
+characters
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.2.0/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.2.0/lib/
+charcode
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/charcode-1.3.1/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/charcode-1.3.1/lib/
+clock
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/clock-1.1.0/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/clock-1.1.0/lib/
+collection
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/collection-1.16.0/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/collection-1.16.0/lib/
+cupertino_icons
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-1.0.5/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-1.0.5/lib/
+fake_async
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/fake_async-1.3.0/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/fake_async-1.3.0/lib/
+flutter_lints
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_lints-1.0.4/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_lints-1.0.4/lib/
+lints
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/lints-1.0.1/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/lints-1.0.1/lib/
+matcher
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.11/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.11/lib/
+material_color_utilities
+2.13
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/material_color_utilities-0.1.4/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/material_color_utilities-0.1.4/lib/
+meta
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/meta-1.7.0/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/meta-1.7.0/lib/
+nested
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/nested-1.0.0/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/nested-1.0.0/lib/
+path
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/path-1.8.1/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/path-1.8.1/lib/
+percent_indicator
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/percent_indicator-4.2.2/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/percent_indicator-4.2.2/lib/
+provider
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/provider-6.0.3/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/provider-6.0.3/lib/
+source_span
+2.14
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/source_span-1.8.2/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/source_span-1.8.2/lib/
+stack_trace
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0/lib/
+stream_channel
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.0/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.0/lib/
+string_scanner
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/string_scanner-1.1.0/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/string_scanner-1.1.0/lib/
+term_glyph
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.2.0/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.2.0/lib/
+test_api
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/test_api-0.4.9/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/test_api-0.4.9/lib/
+vector_math
+2.14
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/vector_math-2.1.2/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/vector_math-2.1.2/lib/
+sky_engine
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/bin/cache/pkg/sky_engine/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/bin/cache/pkg/sky_engine/lib/
+flutter
+2.12
+file:///home/hiltonthallyson/snap/flutter/common/flutter/packages/flutter/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/packages/flutter/lib/
+flutter_test
+2.17
+file:///home/hiltonthallyson/snap/flutter/common/flutter/packages/flutter_test/
+file:///home/hiltonthallyson/snap/flutter/common/flutter/packages/flutter_test/lib/
+2
