+++ catalogo_filmes/lib/screens/catalog_screen.dart
@@ -1,5 +1,5 @@
 import 'dart:convert';
-import 'package:catalogo_filmes/components/my_main_drawer.dart';
+import 'package:catalogo_filmes/components/drawer.dart';
 import 'package:catalogo_filmes/models/movie.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
