+++ catalogo_filmes/lib/utils/app_routes.dart
@@ -1,5 +1,4 @@
 class AppRoutes {
-  static const HOME = '/';
   static const CATALOG = '/catalog';
   static const DETAILS = '/details';
   static const FAVORITES = '/favorites';
